app/requirements.txt
flask
gunicorn
gevent
pillow
pytesseract
